requirements.txt
# absl-py==1.4.0
# attrs==22.2.0
# contourpy==1.0.7
# cycler==0.11.0
# flatbuffers==23.3.3
# fonttools==4.39.2
# kiwisolver==1.4.4
# matplotlib==3.7.1
# mediapipe==0.9.1.0
# MouseInfo==0.1.3
# numpy==1.24.2
# opencv-contrib-python==4.7.0.72
# opencv-python==4.7.0.72
# packaging==23.0
# Pillow==9.4.0
# protobuf==3.20.3
# PyAutoGUI==0.9.53
# PyGetWindow==0.0.9
# PyMsgBox==1.0.9
# pyparsing==3.0.9
# pyperclip==1.8.2
# PyRect==0.2.0
# PyScreeze==0.1.28
# python-dateutil==2.8.2
# pytweening==1.0.4
# six==1.16.0

absl-py
opencv-python
mediapipe
pyautogui